\name{deprecated}
\Rdversion{1.1}
\alias{BamSampler}
\alias{BamSampler-class}
\alias{scanBam,BamSampler-method}
\alias{show,BamSampler-method}
\alias{yieldTabix}
\alias{yieldTabix,TabixFile-method}
\title{Deprecated functions}
\description{Functions listed on this page are no longer supported.}
\details{

  For \code{yieldTabix}, use the \code{yieldSize} argument of
  \code{TabixFiles}. 

  For \code{BamSampler}, use \code{REDUCEsampler} with
  \code{reduceByYield} in the \code{GenomicFiles} package. 
}

\author{

  Martin Morgan <mtmorgan@fhcrc.org>.

}

\keyword{ manip }
