\name{Rsamtools-package}
\Rdversion{1.1}
\alias{Rsamtools-package}
\alias{Rsamtools}
\docType{package}
\title{
Interface to the samtools aligned sequece utilities
}
\description{
}
\details{
See \code{packageDescription('Rsamtools')} for details
}
\author{
Author: Martin Morgan

Maintainer: Biocore Team c/o BioC user list <bioconductor@stat.math.ethz.ch>
}
\references{
  \url{http://samtools.sourceforge.net/}
}
\keyword{ package }
\examples{
packageDescription('Rsamtools')
}
