\name{Alignments-class}
\Rdversion{1.1}

\alias{class:Alignments0}
\alias{Alignments0-class}
\alias{Alignments0}


\title{Alignments0 objects}

\description{
  [TODO]
}

\references{
  \url{http://samtools.sourceforge.net/}
}

\author{
  H. Pages
}

\seealso{
  \link[IRanges]{IRanges-class},
  \link[IRanges]{GappedRanges-class},
  \code{\link[IRanges]{coverage}},
  \link[IRanges]{RleList-class}
}

\examples{
}

\keyword{methods}
\keyword{classes}
