\name{deprecated}
\Rdversion{1.1}
\alias{yieldTabix}
\alias{yieldTabix,TabixFile-method}
\title{Deprecated functions}
\description{Functions listed on this page are no longer supported.}
\details{

  For \code{yieldTabix}, use the \code{yieldSize} argument of
  \code{TabixFiles}. 

}

\author{

  Martin Morgan <mtmorgan@fhcrc.org>.

}

\keyword{ manip }
