\name{Rsamtools-package}
\Rdversion{1.1}
\alias{Rsamtools-package}
\alias{Rsamtools}
\docType{package}
\title{
  'samtools' aligned sequence utilities interface
}
\description{
  This package provides facilities for parsing samtools BAM (binary)
  files representing aligned sequences.
}
\details{

  See \code{packageDescription('Rsamtools')} for package details. A
  useful starting point is the \code{\link{scanBam}} manual page.

}
\author{
Author: Martin Morgan

Maintainer: Biocore Team c/o BioC user list <bioconductor@stat.math.ethz.ch>
}
\references{
  \url{http://samtools.sourceforge.net/}
}
\keyword{ package }
\examples{
packageDescription('Rsamtools')
}
